+++ Src/Classes/Win32/Functions.cs
@@ -199,6 +199,9 @@ public class User32
     public static extern bool IsWindowVisible(nint hWnd);
 
     [DllImport("user32.dll", SetLastError = true)]
+    public static extern bool IsWindow(nint hWnd);
+
+    [DllImport("user32.dll", SetLastError = true)]
     public static extern int AnimateWindow(nint hWnd, uint dwTime, ANIMATEWINDOW dwFlags);
 
     // return hMonitor or the monitor handle